requirements.txt
pymongo[srv]
flask
google-generativeai
python-dotenv
openai
portkey_ai